added vite frontend configurations

+++ backend/controllers/product.controller.js
@@ -49,12 +49,16 @@ export const createProduct = async (req, res) => {
 
     export const deleteProduct = async (req, res) => {
         const {id} = req.params; // Extracting the product ID from the request parameters
+
+        if (!mongoose.Types.ObjectId.isValid(id)) {
+            return res.status(404).json({ success: false, message: 'Invalid product ID' });
+        }
     
         try {
             await Product.findByIdAndDelete(id); // Deleting the product by ID
             res.status(200).json({ success: true, message: 'Product deleted successfully' });
         } catch (error) {
             console.error("Error in deleting product:", error.message);
-            res.status(404).json({ success: false, message: 'Product not found' });
+            res.status(500).json({ success: false, message: 'Server Error' });
         }
     };
